Move atom into its own file
